feat(): Hardening & translation updates.
