+++ app/src/main/res/values-de/strings.xml
@@ -644,6 +644,7 @@
     <string name="achievements_view_all">Alle anzeigen</string>
     <string name="achievements_all_title">Alle Erfolge</string>
     <string name="achievements_unlocked_at">Freigeschaltet: %s</string>
+    <string name="achievements_complete">Alle Erfolge freigeschaltet</string>
     <string name="settings_achievement_notification_position">Position der Erfolgsbenachrichtigung</string>
     <string name="achievement_position_top_left">Oben links</string>
     <string name="achievement_position_top_right">Oben rechts</string>
