+++ app/src/main/res/values-es/strings.xml
@@ -717,6 +717,7 @@
     <string name="achievements_view_all">Ver todos</string>
     <string name="achievements_all_title">Todos los logros</string>
     <string name="achievements_unlocked_at">Desbloqueado: %s</string>
+    <string name="achievements_complete">Todos los logros desbloqueados</string>
     <string name="settings_achievement_notification_position">Posición de notificación de logro</string>
     <string name="achievement_position_top_left">Arriba izquierda</string>
     <string name="achievement_position_top_right">Arriba derecha</string>
