+++ app/src/main/res/values-it/strings.xml
@@ -671,6 +671,7 @@
     <string name="achievements_view_all">Vedi tutti</string>
     <string name="achievements_all_title">Tutti gli obiettivi</string>
     <string name="achievements_unlocked_at">Sbloccato: %s</string>
+    <string name="achievements_complete">Tutti i traguardi sbloccati</string>
     <string name="settings_achievement_notification_position">Posizione notifica obiettivo</string>
     <string name="achievement_position_top_left">In alto a sinistra</string>
     <string name="achievement_position_top_right">In alto a destra</string>
