+++ app/src/main/res/values-pt-rBR/strings.xml
@@ -505,6 +505,7 @@
     <string name="achievements_view_all">Ver todas</string>
     <string name="achievements_all_title">Todas as conquistas</string>
     <string name="achievements_unlocked_at">Desbloqueado em: %s</string>
+    <string name="achievements_complete">Todas as conquistas desbloqueadas</string>
     <string name="settings_achievement_notification_position">Posição da notificação de conquista</string>
     <string name="achievement_position_top_left">Superior esquerdo</string>
     <string name="achievement_position_top_right">Superior direito</string>
