+++ app/src/main/res/values-uk/strings.xml
@@ -666,6 +666,7 @@
     <string name="achievements_view_all">Переглянути всі</string>
     <string name="achievements_all_title">Всі досягнення</string>
     <string name="achievements_unlocked_at">Розблоковано: %s</string>
+    <string name="achievements_complete">Усі досягнення розблоковано</string>
     <string name="settings_achievement_notification_position">Позиція сповіщення про досягнення</string>
     <string name="achievement_position_top_left">Лівий верхній</string>
     <string name="achievement_position_top_right">Правий верхній</string>
