+++ app/src/main/res/values-zh-rTW/strings.xml
@@ -664,6 +664,7 @@
     <string name="achievements_view_all">查看全部</string>
     <string name="achievements_all_title">所有成就</string>
     <string name="achievements_unlocked_at">解鎖時間：%s</string>
+    <string name="achievements_complete">所有成就已解鎖</string>
     <string name="settings_achievement_notification_position">成就通知位置</string>
     <string name="achievement_position_top_left">左上</string>
     <string name="achievement_position_top_right">右上</string>
